Use environment variable for FusionAuth token
This update adds an environment variable for the FusionApp token, replacing the hardcoded token strings in the service and test files. By using `process.env.FUSION_APP_TOKEN`, the security of the token is enhanced and configuration becomes more flexible. The `.env` file is updated to include a sample token for local development.

Signed-off-by: Diego Carhuavilca <diego.carhuavilca@gmail.com>